response, parameterLength에 Unsigned 처리작업(12)

+++ src/main/java/com/devgong/nettyserver/handler/NettyServerHandler.java
@@ -131,7 +131,7 @@ public class NettyServerHandler extends ChannelInboundHandlerAdapter {
                             response.getSn(),
                             LocalDateTime.now(),
                             RequestType.SERVER,
-                            (byte)response.serialize().length + 2,  //4 byte
+                            response.serialize().length + 2,  //4 byte
                             response
                     );
